Splitted the source files. Optimizing flag
This is interesting.
By including all source files into one file,
and adding the option -fwhole-program to the gcc options,
the whole thing is quite more responsive.
So, somehow it seems to me,
the traditional approach of splitting all sources
into source (.c) and header file (.h) is a relict
of times, when every single object file took
at least several seconds to compile.
Nowadays, and my development system is quite old -
this splitting is not needed anymore.
So it seems to me now.
At least for "smaller" projects.